feat: move nvim package management to nix
