+++ userModules/nvim/config/after/plugin/treesitter.lua
@@ -1,0 +1,11 @@
+require 'nvim-treesitter.configs'.setup {
+  sync_install = false,
+  auto_install = false,
+  highlight = {
+    enable = true,
+    additional_vim_regex_highlighting = false,
+  },
+  indent = {
+    enable = true
+  }
+}
