+++ userModules/nvim/config/lua/senyc/init.lua
@@ -1,3 +1,2 @@
 require 'senyc.set' -- Needs to be called first to set global leader
 require 'senyc.keymaps'
-require 'senyc.lazy'
