--- userModules/nvim/config/lua/senyc/plugins/comment.lua
@@ -1,7 +1,0 @@
-return {
-  'numToStr/Comment.nvim',
-  lazy = false,
-  config = function()
-    require 'Comment'.setup()
-  end
-}
