--- userModules/nvim/config/spell/en.utf-8.add
@@ -1,83 +1,0 @@
-png
-j
-peg
-jpg
-mp4
-pdf
-utils
-colorscheme
-onedark
-solargraph
-debounce
-config
-lspconfig
-eslint
-pylsp
-senyc
-ripgrep
-undotree
-tpm
-tmux
-powerline
-UTF-8
-fi
-alias
-clangd
-esac
-keybinds
-cmp
-nvim
-filetype
-filetypes
-allowlist
-denylist
-overriden
-lsp
-treesitter
-NvimTree
-toggleterm
-TelescopePrompt
-lua
-decrementing
-whitelist
-xinitrc
-mason:
-TopLeftCorner
-xl
-px
-scalable
-tailwindcss
-Neovim
-pb
-TypeScript
-npm
-PowerShell
-SharePoint
-CSS
-SCSS
-DaisyUI
-UI
-HTML5
-Vercel
-AWS
-dropdown
-btn
-svg
-luasnip
-aspell
-aspelladd
-github
-dotfiles
-recurence
-returns
-indeterministic
-wikilink
-decrement
-formatoptions
-QuickFix
-macchiato
-first
-prefill
-sessionizer
-sessionizer
-alacritty
